refactor: split 1379-line contributionService into domain modules
Extract three focused modules from the monolithic contributionService.ts:

- submission/validation.ts (196 lines): input validation, sanitization,
  constants (VALIDATION_LIMITS, VALID_AMENITY_KEYS)
- submission/dedup.ts (104 lines): duplicate detection, hash generation,
  recentSubmissions Map with cleanup
- submission/sessionGuard.ts (461 lines): session validation with retry,
  timeout racing with AbortController, exponential backoff

Main file reduced from 1,379 → 634 lines (54% reduction). All public
exports preserved via re-exports so no consumer changes needed. Updated
one test to verify behavior instead of spying on an internal call.
